+++ packages/dataviews/src/dataform-controls/number.tsx
@@ -1,9 +1,0 @@
-/**
- * Internal dependencies
- */
-import type { DataFormControlProps } from '../types';
-import ValidatedNumber from './utils/validated-number';
-
-export default function Number< Item >( props: DataFormControlProps< Item > ) {
-	return <ValidatedNumber { ...props } />;
-}